We centralized our issue tracking on cmangos/issues
All existing issues have been migrated to this centralized repository. As
of now, we will deactivate the issue tracker on this repository and only use
the new central one.
Please note that Pull Requests are still on the repository containing the
code and you'll still post them there, as they are directly associated
with that code.

Thanks to @github for making this possible with
https://github.com/blog/1439-closing-issues-across-repositories

[ci skip]

(based on commit [12426] - e7ecc5f)

Signed-off-by: DasBlub <DasBlub@gmail.com>