Fixed Blink's spell effect teleporting the player to Hillsbrad
Blink was having an issue where it would teleport you to Hillsbrad. This was due to the spell's radius not being properly retrieved from the DBC, which caused an error in the algorithm that the server uses to determine the player's final destination. The error was causing the player's destination coordinates to be set to (0, 0, 0)