Bug fixes for landing from flight and the Blink spell (#89)
* Fixed the player's position being incorrectly set when landing from flight

The server wasn't properly handling the movement opcode CMSG_MOVE_SET_CAN_FLY, which fires when landing or taking off. Because of this, the player's position was being set to (0, 0, 0), which had the effect of making all the surrounding NPCs and gameobjects disappear until moving again.

I've added a placeholder MovementStatusElements sequence for it based on the TrinityCore source code. It's incomplete (the server displays an error about unprocessed tail data), but it is sufficient to fix the bug.

* Fixed Blink's spell effect teleporting the player to Hillsbrad

Blink was having an issue where it would teleport you to Hillsbrad. This was due to the spell's radius not being properly retrieved from the DBC, which caused an error in the algorithm that the server uses to determine the player's final destination. The error was causing the player's destination coordinates to be set to (0, 0, 0)